增加切分字块操作区
