+++ src/router/router.js
@@ -55,15 +55,6 @@ export const page500 = {
     component: resolve => { require(['@/views/error-page/500.vue'], resolve); }
 };
 
-export const confidence = {
-    path: '/confidence',
-    meta: {
-        title: '置信校对'
-    },
-    name: 'zhixin-jiaodui',
-    component: resolve => { require(['@/views/jiaodui/b_confidence.vue'], resolve); }
-};
-
 // export const preview = {
 //     path: '/preview',
 //     name: 'preview',
@@ -273,7 +264,6 @@ export const routers = [
     // preview,
     // locking,
     ...appRouter,
-    confidence,
     page500,
     page403,
     page404
