+++ src/views/main.less
@@ -49,6 +49,7 @@
         z-index: 20;
         box-shadow: 0 2px 1px 1px rgba(100,100,100,.1);
         transition: padding .3s;
+        top: 0px;
     }
     &-breadcrumb{
         padding: 8px 15px 0;
