+++ webpack-stats.json
@@ -1,1 +1,1 @@
-{"status":"done","publicPath":"/dist/","chunks":{"null":[{"name":"8.chunk.js","publicPath":"/dist/8.chunk.js","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/8.chunk.js"},{"name":"8.chunk.js.map","publicPath":"/dist/8.chunk.js.map","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/8.chunk.js.map"}],"main":[{"name":"main.js","publicPath":"/dist/main.js","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/main.js"},{"name":"main.css","publicPath":"/dist/main.css","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/main.css"},{"name":"main.js.map","publicPath":"/dist/main.js.map","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/main.js.map"},{"name":"main.css.map","publicPath":"/dist/main.css.map","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/main.css.map"}],"vendors":[{"name":"vendors.js","publicPath":"/dist/vendors.js","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/vendors.js"},{"name":"vendors.js.map","publicPath":"/dist/vendors.js.map","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/vendors.js.map"}]},"error":"ModuleBuildError","message":"Module build failed: SyntaxError: Unexpected token, expected , (79:29)\n\n  77 |             y_ratio: 1.0,\n  78 |             drag: {\n> 79 |                 current: null;\n     |                              ^\n  80 |                 draggable: false;\n  81 |                 x_ratio: 1.0,\n  82 |                 y_ratio: 1.0,\n"}
+{"status":"done","publicPath":"/dist/","chunks":{"null":[{"name":"10.chunk.js","publicPath":"/dist/10.chunk.js","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/10.chunk.js"},{"name":"10.chunk.js.map","publicPath":"/dist/10.chunk.js.map","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/10.chunk.js.map"}],"main":[{"name":"main.js","publicPath":"/dist/main.js","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/main.js"},{"name":"11.b8ac98bd16e387e5e749.hot-update.js","publicPath":"/dist/11.b8ac98bd16e387e5e749.hot-update.js","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/11.b8ac98bd16e387e5e749.hot-update.js"},{"name":"main.css","publicPath":"/dist/main.css","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/main.css"},{"name":"main.js.map","publicPath":"/dist/main.js.map","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/main.js.map"},{"name":"11.b8ac98bd16e387e5e749.hot-update.js.map","publicPath":"/dist/11.b8ac98bd16e387e5e749.hot-update.js.map","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/11.b8ac98bd16e387e5e749.hot-update.js.map"},{"name":"main.css.map","publicPath":"/dist/main.css.map","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/main.css.map"}],"vendors":[{"name":"vendors.js","publicPath":"/dist/vendors.js","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/vendors.js"},{"name":"vendors.js.map","publicPath":"/dist/vendors.js.map","path":"/Users/zhanghuiyong/Documents/CentralPortal/build/dist/vendors.js.map"}]},"error":"ModuleBuildError","message":"Module build failed: SyntaxError: Unexpected character '#' (56:12)\n\n  54 |             textContent: \"This is TEXT\",\n  55 |             // image_url: \"https://s3.cn-north-1.amazonaws.com.cn/lqcharacters-images/GLZ/S00001/R491/GLZ_S00001_R491_T0024.jpg\",\n> 56 |             #imageUrl: 'http://dist.zhanghuiyong.cn/media/GLZ_S00193_R001_T0012_L01.jpg',\n     |             ^\n  57 |             current: {},\n  58 |             window_height: window.innerHeight,\n  59 |             window_width: window.innerWidth,\n"}
